+++ engines/ags/engine/script/runtime_script_value.cpp
@@ -45,7 +45,7 @@ uint8_t RuntimeScriptValue::ReadByte() const {
 		if (RValue->Type == kScValData) {
 			return *(uint8_t *)(RValue->GetPtrWithOffset() + this->IValue);
 		} else {
-			return RValue->IValue; // get RValue as int
+			return static_cast<uint8_t>(RValue->IValue);
 		}
 	} else if (this->Type == kScValStaticObject || this->Type == kScValStaticArray) {
 		return this->StcMgr->ReadInt8(this->Ptr, this->IValue);
@@ -60,13 +60,13 @@ int16_t RuntimeScriptValue::ReadInt16() const {
 		if (RValue->Type == kScValData) {
 			return *(int16_t *)(RValue->GetPtrWithOffset() + this->IValue);
 		} else {
-			return RValue->IValue; // get RValue as int
+			return static_cast<uint16_t>(RValue->IValue);
 		}
 	} else if (this->Type == kScValGlobalVar) {
 		if (RValue->Type == kScValData) {
 			return Memory::ReadInt16LE(RValue->GetPtrWithOffset() + this->IValue);
 		} else {
-			return RValue->IValue; // get RValue as int
+			return static_cast<uint16_t>(RValue->IValue);
 		}
 	} else if (this->Type == kScValStaticObject || this->Type == kScValStaticArray) {
 		return this->StcMgr->ReadInt16(this->Ptr, this->IValue);
